[Wave] move kernel to templates
