+++ lit_tests/kernel/wave/codegen.py
@@ -7,6 +7,7 @@ import wave_lang.kernel.lang as tkl
 import wave_lang.kernel.wave as tkw
 from wave_lang.kernel.lang.global_symbols import *
 from wave_lang.kernel.wave.compile import WaveCompileOptions, wave_compile
+from wave_lang.kernel.wave.templates.test_kernels import get_broadcast_scaled_add
 from wave_lang.kernel.wave.utils.compile_utils import (
     set_default_compile_config,
 )
@@ -1932,45 +1933,11 @@ def test_broadcast_add():
 
 @run_test
 def test_broadcast_scaled_add():
-    M = tkl.sym.M
-    N = tkl.sym.N
-    ADDRESS_SPACE = tkl.sym.ADDRESS_SPACE
     shape = (256, 256)
-
-    wave_size = 64
-    BLOCK_M = 1
-    BLOCK_N = sympy.Max(sympy.Min(shape[1], 256), wave_size)
-
-    constraints: list[tkw.Constraint] = [
-        tkw.HardwareConstraint(
-            threads_per_wave=wave_size,
-            vector_shapes={M: BLOCK_M, N: BLOCK_N},
-        )
-    ]
-
-    constraints += [tkw.WorkgroupConstraint(M, BLOCK_M, 1)]
-    constraints += [tkw.WorkgroupConstraint(N, BLOCK_N, 0)]
-    constraints += [tkw.WaveConstraint(M, BLOCK_M)]
-    constraints += [tkw.WaveConstraint(N, BLOCK_N)]
-
-    @tkw.wave(constraints)
-    def broadcast_scaled_add(
-        a: tkl.Memory[M, N / 2, ADDRESS_SPACE, tkl.f16],
-        b: tkl.Memory[M, ADDRESS_SPACE, tkl.f16],
-        c: tkl.Memory[M, N / 2, ADDRESS_SPACE, tkl.f16],
-    ):
-        lhs = tkw.read(a)
-        rhs = tkw.read(b)
-        rhs = tkw.broadcast(rhs, [M, N / 2])
-        res = lhs + rhs
-        tkw.write(res, c)
+    broadcast_scaled_add, hyperparams = get_broadcast_scaled_add(shape)
 
     options = WaveCompileOptions(
-        subs={
-            M: shape[0],
-            N: shape[1],
-            ADDRESS_SPACE: tkl.AddressSpace.GLOBAL_MEMORY.value,
-        },
+        subs=hyperparams,
         canonicalize=True,
         use_buffer_ops=True,
         compile_to_mlir=True,
