+++ lit_tests/kernel/wave/scaled_mma.py
@@ -372,77 +372,3 @@ def test_mxfp4_scaled_mma_256x256x256():
     # CHECK-COUNT-64:   amdgpu.scaled_mfma{{.*}} {k = 128 : i32, m = 16 : i32, n = 16 : i32} : f8E8M0FNU, vector<32xf4E2M1FN>, f8E8M0FNU, vector<32xf4E2M1FN>, vector<4xf32>
     # CHECK:            scf.yield
     # CHECK:        }
-
-
-@run_test
-def test_mxfp4_broadcasted_scale_scaled_mma_16x16x128():
-    # Input sizes
-    M = tkl.sym.M
-    N = tkl.sym.N
-    K = tkl.sym.K
-    # Workgroup tile sizes
-    BLOCK_M = tkl.sym.BLOCK_M
-    BLOCK_N = tkl.sym.BLOCK_N
-    BLOCK_K = tkl.sym.BLOCK_K
-    # Address space (for GPU, shared(1) or global(0))
-    ADDRESS_SPACE = tkl.sym.ADDRESS_SPACE
-
-    mfma_variant = ScaledMMAType.F32_16x16x128_F8F6F4
-
-    # Expose user-constraints
-    constraints: list[tkw.Constraint] = [tkw.WorkgroupConstraint(M, BLOCK_M, 0)]
-    constraints += [tkw.WorkgroupConstraint(N, BLOCK_N, 1)]
-    constraints += [tkw.TilingConstraint(K, BLOCK_K)]
-    constraints += [tkw.WaveConstraint(M, BLOCK_M / 2)]
-    constraints += [tkw.WaveConstraint(N, BLOCK_N / 2)]
-
-    constraints += [tkw.HardwareConstraint(threads_per_wave=64, mma_type=mfma_variant)]
-
-    @tkw.wave(constraints)
-    def broadcasted_scale_scaled_mma(
-        a: tkl.Memory[M, K / 2, ADDRESS_SPACE, tkl.i8],
-        a_scale: tkl.Memory[M, ADDRESS_SPACE, tkl.f8e8m0fnu],
-        b: tkl.Memory[N, K / 2, ADDRESS_SPACE, tkl.i8],
-        b_scale: tkl.Memory[N, K / 32, ADDRESS_SPACE, tkl.f8e8m0fnu],
-        c: tkl.Memory[M, N, GLOBAL_ADDRESS_SPACE, tkl.f32],
-    ):
-        a_reg = tkw.read(a)
-        a_reg = tkw.bitcast(a_reg, tkl.f4e2m1fn)
-        a_scale_reg = tkw.read(a_scale)
-        a_scale_reg = tkw.broadcast(a_scale_reg, [M, K / 32])
-        b_reg = tkw.read(b)
-        b_reg = tkw.bitcast(b_reg, tkl.f4e2m1fn)
-        b_scale_reg = tkw.read(b_scale)
-        c_reg = tkl.Register[M, N, tkl.f32](0.0)
-        acc = tkw.scaled_mma(a_reg, a_scale_reg, b_reg, b_scale_reg, c_reg)
-        tkw.write(acc, c)
-
-    hyperparams = {
-        ADDRESS_SPACE: SHARED_ADDRESS_SPACE,
-        BLOCK_M: 32,
-        BLOCK_N: 32,
-        BLOCK_K: 256,
-        M: 32,
-        N: 32,
-        K: 256,
-    }
-    hyperparams.update(get_default_scheduling_params())
-
-    options = WaveCompileOptions(
-        subs=hyperparams,
-        canonicalize=True,
-        backend="rocm",
-        target="gfx950",
-        compile_to_mlir=True,
-    )
-    broadcasted_scale_scaled_mma = wave_compile(options, broadcasted_scale_scaled_mma)
-    print(broadcasted_scale_scaled_mma.asm)
-
-    # This test is important to check that broadcasting on scaled dimension works.
-    # The thing to look out for in this test is the same lhs_scale is being used on the two different mfmas.
-
-    # CHECK-LABEL:  test_mxfp4_broadcasted_scale_scaled_mma_16x16x128
-    # CHECK:   func.func @broadcasted_scale_scaled_mma(%arg0: !stream.binding, %arg1: !stream.binding, %arg2: !stream.binding, %arg3: !stream.binding, %arg4: !stream.binding) attributes {translation_info = #translation} {
-    # CHECK:            %[[LHS_SCALE:.+]] = vector.load {{.*}} : memref<40xf8E8M0FNU, #gpu.address_space<workgroup>>, vector<1xf8E8M0FNU>
-    # CHECK:            %[[LHS_SCALE_EXTRACT:.+]] = vector.extract %[[LHS_SCALE]][0] : f8E8M0FNU from vector<1xf8E8M0FNU>
-    # CHECK-COUNT-2:    amdgpu.scaled_mfma(%[[LHS_SCALE_EXTRACT]][0] * %{{.*}}) * (%{{.*}}[0] * %{{.*}}) + %{{.*}} {k = 128 : i32, m = 16 : i32, n = 16 : i32} : f8E8M0FNU, vector<32xf4E2M1FN>, f8E8M0FNU, vector<32xf4E2M1FN>, vector<4xf32>
