+++ tests/kernel/wave/wave_gemm_mxfp_test.py
@@ -15,6 +15,9 @@ from wave_lang.kernel.wave.utils.torch_utils import (
     device_zeros,
 )
 from wave_lang.kernel.lang.global_symbols import *
+from wave_lang.kernel.wave.templates.test_kernels import (
+    get_broadcasted_scale_gemm_mxfp4,
+)
 from wave_lang.kernel.wave.utils.general_utils import (
     get_default_scheduling_params,
 )
@@ -536,70 +539,16 @@ def testBroadcastedScaleGemmMXFP4(
     shape: tuple[int],
     mfma_variant: ScaledMMAType,
 ):
-    # Input sizes
-    M = tkl.sym.M
-    N = tkl.sym.N
-    K = tkl.sym.K
-    # Workgroup tile sizes
-    BLOCK_M = tkl.sym.BLOCK_M
-    BLOCK_N = tkl.sym.BLOCK_N
-    BLOCK_K = tkl.sym.BLOCK_K
-    # Address space (for GPU, shared(1) or global(0))
-    ADDRESS_SPACE = tkl.sym.ADDRESS_SPACE
-
-    # Expose user-constraints
-    constraints: list[tkw.Constraint] = [tkw.WorkgroupConstraint(M, BLOCK_M, 0)]
-    constraints += [tkw.WorkgroupConstraint(N, BLOCK_N, 1)]
-    constraints += [tkw.TilingConstraint(K, BLOCK_K)]
-    constraints += [tkw.WaveConstraint(M, BLOCK_M / 2)]
-    constraints += [tkw.WaveConstraint(N, BLOCK_N / 2)]
-
-    constraints += [tkw.HardwareConstraint(threads_per_wave=64, mma_type=mfma_variant)]
-
-    @tkw.wave(constraints)
-    def gemm(
-        a: tkl.Memory[M, K / 2, ADDRESS_SPACE, tkl.i8],
-        a_scale: tkl.Memory[M, ADDRESS_SPACE, tkl.i8],
-        b: tkl.Memory[N, K / 2, ADDRESS_SPACE, tkl.i8],
-        b_scale: tkl.Memory[N, K / 32, ADDRESS_SPACE, tkl.i8],
-        c: tkl.Memory[M, N, GLOBAL_ADDRESS_SPACE, tkl.f32],
-    ):
-        c_reg = tkl.Register[M, N, tkl.f32](0.0)
-
-        @tkw.iterate(K, init_args=[c_reg])
-        def repeat(acc: tkl.Register[M, N, tkl.f32]) -> tkl.Register[M, N, tkl.f32]:
-            a_reg = tkw.read(a)
-            a_reg = tkw.bitcast(a_reg, tkl.f4e2m1fn)
-            a_scale_reg = tkw.read(a_scale)
-            a_scale_reg = tkw.broadcast(a_scale_reg, [M, K / 32])
-            a_scale_reg = tkw.bitcast(a_scale_reg, tkl.f8e8m0fnu)
-            b_reg = tkw.read(b)
-            b_reg = tkw.bitcast(b_reg, tkl.f4e2m1fn)
-            b_scale_reg = tkw.read(b_scale)
-            b_scale_reg = tkw.bitcast(b_scale_reg, tkl.f8e8m0fnu)
-            acc = tkw.scaled_mma(a_reg, a_scale_reg, b_reg, b_scale_reg, acc)
-            return acc
-
-        tkw.write(repeat, c)
-
-    hyperparams = {
-        ADDRESS_SPACE: SHARED_ADDRESS_SPACE,
-        BLOCK_M: 32,
-        BLOCK_N: 32,
-        BLOCK_K: 256,
-        M: shape[0],
-        N: shape[1],
-        K: shape[2],
-    }
-    hyperparams.update(get_default_scheduling_params())
-
+    broadcasted_scale_scaled_mma, hyperparams = get_broadcasted_scale_gemm_mxfp4(
+        shape, mfma_variant
+    )
     options = WaveCompileOptions(
         subs=hyperparams,
         canonicalize=True,
         schedule=SchedulingType.NONE,
     )
     options = set_default_run_config(options)
-    gemm = wave_compile(options, gemm)
+    broadcasted_scale_scaled_mma = wave_compile(options, broadcasted_scale_scaled_mma)
 
     x, w, x_scales, w_scales = generate_gemm_afp4wfp4_inputs(shape)
     out = device_zeros(x.shape[0], w.shape[1], dtype=torch.float32)
@@ -608,7 +557,7 @@ def testBroadcastedScaleGemmMXFP4(
     compacted_x_scales = x_scales[:, 0].contiguous()
 
     w_t = w.T.contiguous()
-    gemm(x, compacted_x_scales, w_t, w_scales, out)
+    broadcasted_scale_scaled_mma(x, compacted_x_scales, w_t, w_scales, out)
 
     # Re-broadcast/copy along old dimension
     # [1024] -> [1024, 32]
