+++ wave_lang/kernel/ops/wave_ops.py
@@ -2367,6 +2367,8 @@ class Broadcast(CustomOp, ABC):
     def indexing_dims(self) -> list[IndexSymbol]:
         from ..wave.utils.general_utils import infer_dim
 
+        # infer_dim to handle cases where target_shape has
+        # scaled expressions (i.e K/32 -> K).
         dims = [infer_dim(expr) for expr in self.target_shape]
         return dims
